bio updated and added contact sidebar
