+++ src/styles/_variables.scss
@@ -1,0 +1,4 @@
+$darkTeal: rgb(0, 66, 66);
+$lightTeal: rgb(0, 170, 170);
+$orange: rgb(252, 103, 5);
+$lightOrange: rgb(245, 157, 98);
