+++ src/styles/main.scss
@@ -1,6 +1,7 @@
 // font-family: 'Public Sans', sans-serif;
 // font-family: 'Berkshire Swash', cursive;
 @import "~bootstrap/scss/bootstrap";
+@import './variables';
 
 body {
     background-color: cadetblue;
@@ -15,13 +16,13 @@ body {
     width: 100%;
     height: 100vh;
     font-family: 'Public Sans', sans-serif;
-    color: rgb(252, 103, 5);
-    font-size: 1.5rem;
+    color: $lightOrange;
+    font-size: 18px;
  }
 
  nav {
-    background-color: rgb(0, 66, 66);
-    box-shadow: 5px 5px 5px black;
+    background-color: $darkTeal;
+    box-shadow: 3px 3px 10px black;
 }
 
 h2, h3, .navbar-brand {
@@ -50,15 +51,23 @@ h2, h3, .navbar-brand {
 
  a:hover {
     color: rgb(252, 103, 5);
+    text-decoration: none;
  }
 
 #page-container {
     padding: 5em 4em 0 4em;
-    background-color: rgba(58, 58, 58, 0.7);
+    background-color: transparentize($color: $darkTeal, $amount: 0.3);
+    box-shadow: 0 0 55px black;
+    min-height: 100vh;
 }
 
 #projectsPage {
     min-height: 550px;
+    display: none;
+}
+
+#technologiesPage {
+    display: none;
 }
 
 #mainDiv {
@@ -67,6 +76,91 @@ h2, h3, .navbar-brand {
 }
 
 .project-card {
-    background-color: rgba(58, 58, 58, 0.75);;
+    background-color: $darkTeal;
     box-shadow: 5px 5px 10px black;
+    color: $lightTeal;
+    border-radius: 0;
+    font-size: 16px;
+    padding: 0;
+
+    p {
+        padding: 8px;
+        margin: 0;
+    }
 }
+
+.card-footer {
+    padding: 0;
+    margin: 0;
+}
+
+.btn-block {
+    color: $lightOrange;
+    border-radius: 0;
+
+    &:hover {
+    background-color: $lightTeal;
+
+        &a {
+            color: $darkTeal;
+        }
+    }
+}
+
+.btn-block + .btn-block {
+    margin-top: 0;
+}
+
+.card-title {
+    text-align: center;
+    margin: 0.5rem 0 0.5rem 0;
+}
+
+// ----Social Icon Sidebar------
+/* Fixed/sticky icon bar (vertically aligned 50% from the top of the screen) */
+.icon-bar {
+    position: fixed;
+    top: 40%;
+    left: 0;
+    -webkit-transform: translateY(-50%);
+    -ms-transform: translateY(-50%);
+    transform: translateY(-50%);
+    box-shadow: 3px 3px 10px black;
+  }
+  
+  /* Style the icon bar links */
+  .icon-bar a {
+    display: block;
+    text-align: center;
+    padding: 16px;
+    transition: all 0.3s ease;
+    color: white;
+    font-size: 20px;
+  }
+  
+  /* Style the social media icons with color, if you want */
+    .icon-bar a:hover {
+    background-color: $orange;
+  }
+
+  .email {
+    background: $darkTeal;
+    color: white;
+  }
+  
+  .linkedin {
+    background: $lightTeal;
+    color: white;
+  }
+  
+  .gitHub {
+      background: $darkTeal;
+      color: white;
+    }
+    
+    .twitter {
+      background: $lightTeal;
+      color: white;
+    }
+
+// ----End Social Icon Sidebar---
